+++ packages/core/ios/RNSentry.mm
@@ -758,18 +758,22 @@ RCT_EXPORT_METHOD(setTag : (NSString *)key value : (NSString *)value)
 
 RCT_EXPORT_METHOD(setAttribute : (NSString *)key value : (NSString *)value)
 {
-    // TODO(alwx): This is not implemented in sentry-cocoa yet
-    /*[SentrySDKWrapper
-        configureScope:^(SentryScope *_Nonnull scope) { [scope setAttribute:value forKey:key]; }];*/
+    [SentrySDKWrapper configureScope:^(
+        SentryScope *_Nonnull scope) { [scope setAttributeValue:value forKey:key]; }];
 }
 
 RCT_EXPORT_METHOD(setAttributes : (NSDictionary *)attributes)
 {
-    // TODO(alwx): This is not implemented in sentry-cocoa yet
-    /*[SentrySDKWrapper configureScope:^(SentryScope *_Nonnull scope) {
-        [attributes enumerateKeysAndObjectsUsingBlock:^(
-            NSString *key, NSString *value, BOOL *stop) { [scope setAttribute:value forKey:key]; }];
-    }];*/
+    [SentrySDKWrapper configureScope:^(SentryScope *_Nonnull scope) {
+        [attributes enumerateKeysAndObjectsUsingBlock:^(NSString *key, NSString *value,
+            BOOL *stop) { [scope setAttributeValue:value forKey:key]; }];
+    }];
+}
+
+RCT_EXPORT_METHOD(removeAttribute : (NSString *)key)
+{
+    [SentrySDKWrapper
+        configureScope:^(SentryScope *_Nonnull scope) { [scope removeAttributeForKey:key]; }];
 }
 
 RCT_EXPORT_METHOD(crash) { [SentrySDKWrapper crash]; }